+++ pykechain/models/property_attachment.py
@@ -72,7 +72,7 @@ class AttachmentProperty(Property, AttachmentProperty2):
         """
         return self._download().json()
 
-    def upload(self, data: Text, **kwargs: Any) -> None:
+    def upload(self, data: Any, **kwargs: Any) -> None:
         """Upload a file to the attachment property.
 
         When providing a :class:`matplotlib.figure.Figure` object as data, the figure is uploaded as PNG.
@@ -137,7 +137,7 @@ class AttachmentProperty(Property, AttachmentProperty2):
 
         response = self._client._request('GET', url)
 
-        if response.status_code != requests.codes.ok:
+        if response.status_code != requests.codes.ok:  # pragma: no cover
             raise APIError("Could not download property value.", response=response)
 
         return response
@@ -150,5 +150,5 @@ class AttachmentProperty(Property, AttachmentProperty2):
                                          data={"part": self._json_data['part_id']},
                                          files={"attachment": data})
 
-        if response.status_code != requests.codes.ok:
+        if response.status_code != requests.codes.ok:  # pragma: no cover
             raise APIError("Could not upload attachment", response=response)