+++ CHANGELOG.rst
@@ -1,13 +1,31 @@
 Change Log
 ==========
 
-v3.8.3-rc3 (UNRELEASED)
------------------------
+v3.8.1 (04SEP20)
+----------------
+
+* :bug: Added `title_visible` property to `Widget` class for the widget's title shown in KE-chain, deprecating the `default_title` mechanism when creating widgets. The default title of a widget is not dependent on front-end and is not stored in the widget.
+* :bug: Set and Update of widget associations now handles optional `part_instance_id` and/or `parent_part_instance_id` inputs.
+* :bug: `WidgetsManager` and `PartSet` no longer implement `Iterable` as an "iterator", making it possible to loop over the Widgets/Parts multiple times.
+* :bug: `add_signature_widget()` method of the `WidgetManager` class now creates an editable signature widget by default. The new input argument `editable` can be set to False to create a viewable widget.
+
+* :+1: Added `update_activities` method to `Client` to update activities in bulk.
+* :+1: `WidgetsManager` is now stored in its `Activity` object for lazy retrieval, while the `WidgetsManager` no explicitly stores a reference to its `Activity`.
+* :+1: Available `Part` options of a pre-filtered multi-reference properties are now filtered when using the `choices()` method on the `MultiReferenceProperty`.
+* :+1: Added `model_id` attribute to `Part` class.
+* :+1: Added `count_instances()` method to the `Part` class, to retrieve the number of Part instances of a Part model.
+* :+1: Added `get_landing_page_url()` method to the `Scope` class, to retrieve the (relative) URL of landing page for that scope. Append it to the client's API root for a full URL.
+* :+1: Added `LanguageCodes` enum class to enumerate the available Language options for user profiles.
+* :+1: Added `value_ids()` method to `_ReferenceProperty` class, returning a list of UUIDs instead of Pykechain objects.
+* :+1: Added lazy retrieval in `parent()` method of `TreeObject`, `Part` and `Activity` classes. Retrieving children or populating descendants also sets all known parent objects.
+
+v3.8.1-rc3 (03SEP20)
+--------------------
 
 * :+1: Inverted the inheritance hierarchy of Class2 classes, allowing for type-checking via `isinstance()`. However, creating objects from these classes is no longer supported.
 
-v3.8.1-rc2 (UNRELEASED)
------------------------
+v3.8.1-rc2 (19AUG20)
+--------------------
 
 * :+1: Added `set_associations` and `remove_associations` method to the `Widget` class (#827)
 * :+1: Renamed activity clone API endpoint is now supported. We now support the cloning and renaming of Parts as well as cloning the activities. (#805)